add tiny changes, mu100 still not finished

+++ src/modules/config.py
@@ -4,7 +4,7 @@ class MNISTBagsConfig(SimpleNamespace):
     seed: int
     positive_num: int
     mean_bag_size: int
-    var_bag_size: int
+    var_bag_size: float
     num_bags: int
     train: bool
